pyproject.toml
[project]
name = "cvat-custom-model-orchestrator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.100.0",
    "numpy>=1.24.0",
    "opencv-python-headless>=4.8.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "ultralytics>=8.0.0",
    "uvicorn>=0.22.0",
]

[dependency-groups]
dev = [
    "httpx>=0.24.0",
    "pytest>=7.4.0",
]
